Working Matrix Version!!!
After analysing dEdI for the layer which the input neurons feed into, it was revealed that the matrix version had 0.94's while the normal version had all zeros. You would then expect that it was the old version which was wrong all along, even though it performed better than matrix.
This turned out not to be the case, as the dEdI's we're actually so close to zero that they appeared so, and that it was multiplying by the activation function derivative which lowered the values so much, and which was missing from the matrix version.
The reason for such a low derivative was because the summed input of each neuron was so large due to all the negative one's in the image data (representing white), that the slope of tanh(inputSum) was essentially zero for every neuron.
This can be fixed probably by either replacing the activation function with something that does not flatten out so much at the higher values, or by scaling down the summed input for each neuron to take into account the number of neurons feeding into it on each layer.