A pretty bad tictactoe bot evolution
A population of bots are selected for their ability to play against a random bot. They can reach ~90-95% win rate playing first, and ~70-73% win rate playing second. This is technically better than a completely random bot which would win 58% of the time playing first, and 28% of the time playing second.
Work still needs to be done to allow for faster training, and also the ability to converge at a more global maximum than currently.